+++ apps/decodex/src/state/store/leases/acquisition.rs
@@ -1,0 +1,124 @@
+use std::{
+	collections::HashSet,
+	fs::{self, OpenOptions, TryLockError},
+};
+
+use crate::{
+	prelude::{Result, eyre},
+	state::{
+		runtime_records::GuardRetention,
+		store::{self, DispatchSlotGuard, IssueClaimGuard, IssueLease, StateStore},
+	},
+};
+
+impl StateStore {
+	/// Try to acquire one issue claim plus one shared dispatch slot for one issue.
+	pub fn try_acquire_lease(
+		&self,
+		project_id: &str,
+		issue_id: &str,
+		run_id: &str,
+		issue_state: &str,
+	) -> Result<bool> {
+		let mut state = self.lock()?;
+
+		if state.leases.values().any(|lease| lease.issue_id == issue_id) {
+			return Ok(false);
+		}
+
+		if let Some(dispatch_slot_config) = state.dispatch_slot_configs.get(project_id).cloned() {
+			fs::create_dir_all(&dispatch_slot_config.root)?;
+
+			let _coordinator = store::acquire_shared_lock_coordinator(&dispatch_slot_config.root)?;
+			let issue_claim_lock_path =
+				store::issue_claim_lock_path(&dispatch_slot_config.root, issue_id);
+			let issue_claim_lock_file = OpenOptions::new()
+				.read(true)
+				.write(true)
+				.create(true)
+				.truncate(false)
+				.open(&issue_claim_lock_path)?;
+
+			match issue_claim_lock_file.try_lock() {
+				Ok(()) => {},
+				Err(TryLockError::WouldBlock) => return Ok(false),
+				Err(TryLockError::Error(error)) => return Err(error.into()),
+			}
+
+			let mut issue_claim_guard = IssueClaimGuard {
+				lock_path: issue_claim_lock_path,
+				lock_file: issue_claim_lock_file,
+				retention: GuardRetention::Local,
+			};
+
+			store::write_issue_claim_record(
+				&mut issue_claim_guard.lock_file,
+				project_id,
+				issue_id,
+				run_id,
+				issue_state,
+			)?;
+
+			let held_slot_indexes = state
+				.dispatch_slot_guards
+				.values()
+				.filter(|guard| guard.project_id == project_id)
+				.map(|guard| guard.slot_index)
+				.collect::<HashSet<_>>();
+			let mut slot_index = 0;
+			let dispatch_slot_guard = loop {
+				if held_slot_indexes.contains(&slot_index) {
+					slot_index = slot_index
+						.checked_add(1)
+						.ok_or_else(|| eyre::eyre!("dispatch slot index overflowed usize"))?;
+
+					continue;
+				}
+
+				let dispatch_slot_lock_path =
+					store::dispatch_slot_lock_path(&dispatch_slot_config.root, slot_index);
+				let lock_file = OpenOptions::new()
+					.read(true)
+					.write(true)
+					.create(true)
+					.truncate(false)
+					.open(&dispatch_slot_lock_path)?;
+
+				match lock_file.try_lock() {
+					Ok(()) => {
+						break DispatchSlotGuard {
+							project_id: project_id.to_owned(),
+							slot_index,
+							lock_path: dispatch_slot_lock_path,
+							lock_file,
+							retention: GuardRetention::Local,
+						};
+					},
+					Err(TryLockError::WouldBlock) => {},
+					Err(TryLockError::Error(error)) => return Err(error.into()),
+				}
+
+				slot_index = slot_index
+					.checked_add(1)
+					.ok_or_else(|| eyre::eyre!("dispatch slot index overflowed usize"))?;
+			};
+
+			state.issue_claim_guards.insert(issue_id.to_owned(), issue_claim_guard);
+			state.dispatch_slot_guards.insert(issue_id.to_owned(), dispatch_slot_guard);
+		}
+
+		state.leases.insert(
+			issue_id.to_owned(),
+			IssueLease {
+				project_id: project_id.to_owned(),
+				issue_id: issue_id.to_owned(),
+				run_id: run_id.to_owned(),
+				issue_state: issue_state.to_owned(),
+			},
+		);
+		state.remember_run_project(project_id, issue_id, Some(run_id));
+		self.persist_runtime_state_locked(&state)?;
+
+		Ok(true)
+	}
+}
